requirements.txt
readability-lxml==0.7.1
requests==2.23.0
spacy>=2.0.16
boilerpy3~=1.0.2